+++ web/packages/new/photos/services/settings.ts
@@ -6,6 +6,7 @@
 
 import { localUser } from "@/base/local-user";
 import log from "@/base/log";
+import { customAPIOrigin } from "@/base/origins";
 import { nullToUndefined } from "@/utils/transform";
 import { z } from "zod";
 import { fetchFeatureFlags, updateRemoteFlag } from "./remote-store";
@@ -193,3 +194,40 @@ export const updateMapEnabled = async (isEnabled: boolean) => {
     await updateRemoteFlag("mapEnabled", isEnabled);
     return syncSettings();
 };
+
+/**
+ * Return true to disable the upload of files via Cloudflare Workers.
+ *
+ * These workers were introduced as a way of make file uploads faster:
+ * https://ente.io/blog/tech/making-uploads-faster/
+ *
+ * By default, that's the route we take. However, during development or when
+ * self-hosting it can be convenient to turn this flag on to directly upload to
+ * the S3-compatible URLs returned by the ente API.
+ *
+ * Note the double negative (Enhancement: maybe remove the double negative,
+ * rename this to say getUseDirectUpload).
+ */
+export async function getDisableCFUploadProxyFlag(): Promise<boolean> {
+    // If a custom origin is set, that means we're not running a production
+    // deployment (maybe we're running locally, or being self-hosted).
+    //
+    // In such cases, disable the Cloudflare upload proxy (which won't work for
+    // self-hosters), and instead just directly use the upload URLs that museum
+    // gives us.
+    if (await customAPIOrigin()) return true;
+
+    try {
+        const featureFlags = (
+            await fetch("https://static.ente.io/feature_flags.json")
+        ).json() as GetFeatureFlagResponse;
+        return featureFlags.disableCFUploadProxy ?? false;
+    } catch (e) {
+        log.error("failed to get feature flags", e);
+        return false;
+    }
+}
+
+export interface GetFeatureFlagResponse {
+    disableCFUploadProxy?: boolean;
+}
